Get rid of math.h

+++ polyweb.cpp
@@ -1,11 +1,11 @@
 #include "polyweb.hpp"
 #include <algorithm>
 #include <bitset>
+#include <cmath>
 #include <codecvt>
 #include <iomanip>
 #include <iterator>
 #include <locale>
-#include <math.h>
 #include <sstream>
 #include <string.h>
 #include <wchar.h>
@@ -15,7 +15,7 @@
 
 namespace pw {
     // f(x) = 2 * log2(x) + x + 4
-    tp::ThreadPool threadpool(roundf(2.f * log2f(std::thread::hardware_concurrency()) + std::thread::hardware_concurrency() + 4.f));
+    tp::ThreadPool threadpool(std::round(2. * std::log2(std::thread::hardware_concurrency()) + std::thread::hardware_concurrency() + 4.));
 
     namespace detail {
         thread_local int last_error = PW_ESUCCESS;
